DBOutputter.cpp: workaround for compilation failure c.f. issue #40
FossilOrigin-Name: d11aef48663b5a97b3cd9499466e81a91fc68f067fb22ec377bd34cf173ed965

+++ qcdb/src/DBOutputter.cpp
@@ -36,7 +36,19 @@ namespace qcdb {
 				) ;
 			}
 		
-			bool match_rsid = (std::find( fields.begin(), fields.end(), CompareFields::eRSID ) != fields.end()) || (std::find( fields.begin(), fields.end(), CompareFields::eIDs ) != fields.end() ) ;
+			bool match_rsid = (
+				std::find(
+					fields.begin(),
+					fields.end(),
+					static_cast<int>(CompareFields::eRSID)
+				) != fields.end()
+			) || (
+				std::find(
+					fields.begin(),
+					fields.end(),
+					static_cast<int>(CompareFields::eIDs)
+				) != fields.end()
+			) ;
 			return match_rsid ;
 		}
 	}
